requirements.txt
# requirements.txt
PyMuPDF>=1.19.0  # For PDF processing (fitz)
Pillow>=8.0.0    # For image handling (PIL)